rename: jwt 토큰 파일 타입 오타 수정
