+++ src/apps/server/auth/types/jwt-token.type.ts
