+++ src/apps/server/experiences/interface/experience-repository.interface.ts
@@ -1,7 +1,7 @@
 import { Experience, ExperienceCapability, ExperienceInfo, Prisma } from '@prisma/client';
 import { UpdateExperienceReqDto } from '../dto/req/updateExperience.dto';
 import { ExperienceSelect } from './experience-select.interface';
-import { UserJwtToken } from '🔥apps/server/auth/types/jwt-tokwn.type';
+import { UserJwtToken } from '🔥apps/server/auth/types/jwt-token.type';
 
 export interface ExperienceTransactionInterface {
   createExperienceInfo(body: UpdateExperienceReqDto, user: UserJwtToken): Promise<[Experience, ExperienceInfo]>;
